Adjust performance timing
doExpiredTasks should return the actual time it takes to complete.
Clamp the result to give the network phase 50-150ms. Move the watchdog
update call to after reporting performance statistics. This way the tick
will complete _then_ either the watchdog kills the process for taking too
long, we sleep for the remainder of the tick, or just loop, reporting
degraded performance if it's in the acceptable range (<100ms).